Implement first attempt at scheduled API requests
Build function getAllMatches, which schedules a request to the Riot API
for a list of match IDs and corresponding requests to get match data
for all IDs received. For debugging purposes, this function currently
only pulls 3 matches for a given PUUID but will change in the future
to get all matches available.

Also create an exportMatch function to store received match data in
JSON format for debugging purposes.