Implement basic match data processing
Create the classes WinrateTable and ChampData, which enable storage of
match data in the desired output format. WinrateTable represents the
table which will be printed in response to a user's request to retrieve
data for a given League of Legends account. As such, it contains an
object consisting of (key:value) pairs in the format (champion name:
ChampData object).

A ChampData object represents a row in the data
table, which will contain aggregated game data for a particular League
of Legends champion across all ARAM games the requested player has
participated in (dating back as far as July 2021, when Riot started
recording game timestamps). Fields represented in ChampData include
wins, losses, total games, and winrate, organized by when the champion
appeared on the requested player's team, on the enemy team, or on
either team.

WinrateTable takes a string array in its constructor representing the
names of all League of Legends characters to aid with table
initialization. For this purpose, the helper function getChampNames
accesses this information from DataDragon, static game data which Riot
Games publishes with each new patch.