Connect backend to frontend search function
The web server has nginx configured to supply a generic webpage
summoner.html to any URL for which an existing file does not exist.
Summoner.html interprets the path of the URL as a summoner name to
query in the backend, and displays the associated data table once it is
received.