Begin MongoDB integration
Start implementation of MongoDB backend for caching player stats.
Write the getProfile method, which accepts an account name as an
argument and returns their current statistics table. If the account
already exists in the database, return the stored table; otherwise,
create a new table from scratch and add it to the database.

Previously, a summoner's matches were analyzed atomically with the
retrieval of their match IDs. Since this database implementation will
seek to split match ID retrieval and match analysis into distinct
operations, WinrateTable now contains a new field, unloggedGames, which
represents a player's unlogged match IDs in a stack. This will enable
the future creation of a new method, updateProfile, which will analyze
matches from the stack to provide small but fast updates to the
frontend via long polling.