Add incremental profile updates for long polling
Write updateProfile, a function which accepts a summoner name as a
parameter and returns an updated table that includes a small number of
newly analyzed games (currently 3). updateProfile also updates the
summoner's WinrateTable in the database accordingly. The frontend will
use long polling to make repeated HTTP requests to a backend endpoint
that will invoke updateProfile to slowly add new data to a given
summoner's profile in real time.